+++ sdk/src/main/java/com/horizen/transaction/FeePaymentsTransaction.java
@@ -5,6 +5,8 @@ import com.horizen.proposition.PublicKey25519Proposition;
 import com.horizen.box.ZenBox;
 import com.horizen.transaction.exception.TransactionSemanticValidityException;
 import com.horizen.transaction.incompatibilitychecker.MempoolIncompatibleTransactionIncompatibilityChecker;
+import com.horizen.transaction.incompatibilitychecker.TransactionIncompatibilityChecker;
+import com.horizen.transaction.serializers.FeePaymentsTransactionSerializer;
 import scala.Array;
 import sparkz.core.serialization.SparkzSerializer;
