+++ sdk/src/main/java/com/horizen/transaction/KeyRotationTransaction.java
@@ -15,6 +15,7 @@ import com.horizen.secret.PrivateKey25519;
 import com.horizen.secret.Secret;
 import com.horizen.transaction.exception.TransactionSemanticValidityException;
 import com.horizen.transaction.incompatibilitychecker.KeyRotationTransactionIncompatibilityChecker;
+import com.horizen.transaction.incompatibilitychecker.TransactionIncompatibilityChecker;
 import com.horizen.transaction.serializers.KeyRotationTransactionSerializer;
 import com.horizen.utils.Pair;
 import scala.Array;
