+++ sdk/src/main/java/com/horizen/transaction/MC2SCAggregatedTransaction.java
@@ -9,7 +9,9 @@ import com.horizen.proposition.Proposition;
 import com.horizen.serialization.Views;
 import com.horizen.transaction.exception.TransactionSemanticValidityException;
 import com.horizen.transaction.incompatibilitychecker.MempoolIncompatibleTransactionIncompatibilityChecker;
+import com.horizen.transaction.incompatibilitychecker.TransactionIncompatibilityChecker;
 import com.horizen.transaction.mainchain.*;
+import com.horizen.transaction.serializers.MC2SCAggregatedTransactionSerializer;
 import com.horizen.utils.*;
 import scala.Array;
 import scorex.util.encode.Base16;
