+++ sdk/src/main/java/com/horizen/transaction/incompatibilitychecker/MempoolIncompatibleTransactionIncompatibilityChecker.java
@@ -1,7 +1,6 @@
 package com.horizen.transaction.incompatibilitychecker;
 
 import com.horizen.transaction.BoxTransaction;
-import com.horizen.transaction.TransactionIncompatibilityChecker;
 
 import java.util.List;
