+++ sdk/src/main/java/com/horizen/transaction/serializers/FeePaymentsTransactionSerializer.java
@@ -1,7 +1,9 @@
-package com.horizen.transaction;
+package com.horizen.transaction.serializers;
 
 import com.horizen.box.ZenBox;
 import com.horizen.box.ZenBoxSerializer;
+import com.horizen.transaction.FeePaymentsTransaction;
+import com.horizen.transaction.TransactionSerializer;
 import com.horizen.utils.ListSerializer;
 import scorex.util.serialization.Reader;
 import scorex.util.serialization.Writer;
