+++ sdk/src/main/scala/com/horizen/block/MainchainBlockReferenceData.scala
@@ -5,7 +5,8 @@ import com.horizen.block.SidechainCreationVersions.SidechainCreationVersion
 import com.horizen.cryptolibprovider.FieldElementUtils
 import com.horizen.serialization.Views
 import com.horizen.transaction.mainchain.{ForwardTransfer, SidechainCreation}
-import com.horizen.transaction.{MC2SCAggregatedTransaction, MC2SCAggregatedTransactionSerializer}
+import com.horizen.transaction.MC2SCAggregatedTransaction
+import com.horizen.transaction.serializers.MC2SCAggregatedTransactionSerializer
 import sparkz.core.serialization.{BytesSerializable, SparkzSerializer}
 import scorex.util.serialization.{Reader, Writer}
