+++ sdk/src/main/scala/com/horizen/companion/SidechainTransactionsCompanion.scala
@@ -6,7 +6,7 @@ import com.horizen.transaction.CoreTransactionsIdsEnum._
 import java.util.{HashMap => JHashMap}
 import java.lang.{Byte => JByte}
 import com.horizen.SidechainTypes
-import com.horizen.transaction.serializers.KeyRotationTransactionSerializer
+import com.horizen.transaction.serializers.{KeyRotationTransactionSerializer, MC2SCAggregatedTransactionSerializer, OpenStakeTransactionSerializer, SidechainCoreTransactionSerializer}
 import com.horizen.utils.DynamicTypedSerializer
